requirements.txt
# Use this file to install the required packages:
# pip install -r requirements.txt

pygame-ce~=2.5.5
moderngl~=5.12.0
Pillow==11.2.1
numpy~=2.3.1